230526_PJT/finalpjtback/requirements.txt
asgiref==3.6.0
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.18
django-allauth==0.54.0
django-bootstrap-v5==1.0.11
django-cors-headers==4.0.0
django-extensions==3.2.1
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.9.0
idna==3.4
ipython-genutils==0.2.0
oauthlib==3.2.2
pickleshare==0.7.5
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.1
text-unidecode==1.3
toposort==1.10
traitlets==5.0.5
urllib3==2.0.2
wcwidth==0.2.5
